Add Candidat class
